Changing timing to be longs instead of ints
Matches upstream WPILib. Allows time greater then 72 minutes.

+++ HAL/SimulatorHAL/HAL.cs
@@ -413,10 +413,10 @@ namespace HAL.SimulatorHAL
         /// <param name="status"></param>
         /// <returns></returns>
         [CalledSimFunction]
-        public static uint getFPGATime(ref int status)
+        public static ulong getFPGATime(ref int status)
         {
             status = 0;
-            return (uint)SimHooks.GetFPGATime();
+            return (ulong)SimHooks.GetFPGATime();
         }
 
         /// <summary>
